change to rate number

+++ src/gradient_descent/Friend.java
@@ -1,20 +1,20 @@
 package gradient_descent;
 
 public class Friend {
-	public Friend(String name, int vote) {
-		super();
-		this.name = name;
-		this.vote = vote;
-	}
+    public Friend(String name, double vote) {
+        super();
+        this.name = name;
+        this.vote = vote;
+    }
 
-	private String name;
-	private int vote;
+    private String name;
+    private double vote;
 
-	public String getName() {
-		return name;
-	}
+    public String getName() {
+        return name;
+    }
 
-	public int getVote() {
-		return vote;
-	}
+    public double getVote() {
+        return vote;
+    }
 }
