+++ src/gradient_descent/MovieInfor.java
@@ -4,20 +4,20 @@ import java.util.List;
 
 public class MovieInfor {
 
-	private int isLike;
-	private List<Friend> friendVoteList;
+    private double isLike;
+    private List<Friend> friendVoteList;
 
-	public MovieInfor(int isLike, List<Friend> friendVoteList) {
-		super();
-		this.isLike = isLike;
-		this.friendVoteList = friendVoteList;
-	}
+    public MovieInfor(double isLike, List<Friend> friendVoteList) {
+        super();
+        this.isLike = isLike;
+        this.friendVoteList = friendVoteList;
+    }
 
-	public int isLike() {
-		return isLike;
-	}
+    public double isLike() {
+        return isLike;
+    }
 
-	public List<Friend> getFriendVoteList() {
-		return friendVoteList;
-	}
+    public List<Friend> getFriendVoteList() {
+        return friendVoteList;
+    }
 }
